fix(modules): re-export types, conform to *Schema convention (#795)
Pre-typebox, we exported all interfaces & types, actually because this was
required for schema generation.  Now that we don't need that anymore,
whether or not we should do this in the longterm is a conversation that
deserves merit, however, since some users make sure of these exports, we
need to export them all to avoid it being a breaking change.

In `quoteSummmary-iface.ts`, renamed the schemas to have a *Schema suffix
and exported types without that suffix, like in other modules.  Similar
small change like that in `quoteSummary.ts` too.

cc: @eddie-atkinson