fix(types): use StaticDecode vs Static to infer types (#797)
With Typebox, `Static` is used to infer types from schema, however,
`StaticDecode` is needed for Transform types which have a `Decode()`
function.  This is important also for schemas composed of other
schemas that make use of transform types.

We already use `StaticDecode` in `validateAndCoerceTypes()` which
is the most important place and filters down to regular use of the
library.  But now that we are exporting all types again (per previous
commit), this is important to make sure that users who are importing
types directly get the correct type.  And I think solved some edge
cases where certain types will still coming through incorrectly.

cc: @eddie-atkinson