feat(validation): use our own vs 3rd party validation
AJV uses `new Function()` which is disabled in some environments,
e.g. Cloudflare workers.  It's possible to precompile the functions
at build time* but this gets very complicated.

* We started this approach and you can see scripts/schema-compile.ts.
It got complicated with the custom yahooFinanceTypes.  Also, the
resulting file was 1.5mb even before dealing with that which was
a shame.

Since we only use a tiny subset of the entire JSONSchema spec, we
really can handle this in-house.